@@ -24,52 +24,57 @@
     "package:linux": " yarn run release:build && electron-packager . slippr --app-version=0.1.0 --build-version=0.1.0 --platform=linux  --arch=all --icon=./assets/images/icons/icns/1024x1024.png  --out=./packages",
     "package:darwin": "yarn run release:build && electron-packager . slippr --app-version=0.1.0 --build-version=0.1.0 --platform=darwin --arch=all --icon=./assets/images/icons/icns/1024x1024.icns --out=./packages",
     "package:win": "   yarn run release:build && electron-packager . slippr --app-version=0.1.0 --build-version=0.1.0 --platform=win32  --arch=all --icon=./assets/images/icons/ico/512x512.ico     --out=./packages",
-    "zip": "node tasks/zip.js 0.1.0"
+    "zip": "node tasks/zip.js 0.1.0",
+    "postinstall": "npm rebuild node-sass --force"
   },
   "dependencies": {
     "babel": "^6.23.0",
     "classnames": "^2.2.5",
-    "electron": "^1.6.2",
+    "electron": "^1.6.10",
     "emojify.js": "^1.1.0",
-    "highlight.js": "^9.10.0",
+    "highlight.js": "^9.12.0",
+    "history": "^4.6.1",
     "immutable": "^3.8.1",
     "jquery": "^3.2.1",
     "jquery-textcomplete": "^1.8.0",
     "markdown-it": "^8.3.1",
     "markdown-it-link-target": "^1.0.2",
-    "node-sass": "^4.5.1",
-    "photon": "https://github.com/connors/photon",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
-    "react-redux": "^5.0.3",
-    "react-router": "^3.0.0",
+    "node-sass": "^4.5.3",
+    "photon": "https://github.com/connors/photon.git",
+    "prop-types": "^15.5.10",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-redux": "^5.0.5",
+    "react-router": "^4.1.1",
+    "react-router-dom": "^4.1.1",
     "redux": "^3.6.0",
-    "redux-logger": "^2.10.2",
+    "redux-logger": "^3.0.6",
     "redux-thunk": "^2.2.0"
   },
   "devDependencies": {
-    "autoprefixer": "^6.7.7",
-    "babel-core": "^6.24.0",
-    "babel-loader": "^6.4.1",
+    "autoprefixer": "^7.1.1",
+    "babel-core": "^6.25.0",
+    "babel-loader": "^7.0.0",
     "babel-polyfill": "^6.23.0",
-    "babel-preset-es2015": "^6.24.0",
-    "babel-preset-react": "^6.23.0",
-    "babel-preset-react-app": "^2.2.0",
-    "babel-preset-stage-3": "^6.22.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-react-app": "^3.0.0",
+    "babel-preset-stage-3": "^6.24.1",
     "ejs": "^2.5.6",
-    "electron-packager": "^8.6.0",
-    "eslint": "^3.18.0",
-    "eslint-config-standard": "^7.1.0",
-    "eslint-loader": "^1.6.0",
-    "eslint-plugin-import": "^2.0.1",
-    "eslint-plugin-jsx-a11y": "^4.0.0",
+    "electron-packager": "^8.7.1",
+    "eslint": "^3.19.0",
+    "eslint-config-standard": "^10.2.1",
+    "eslint-loader": "^1.7.1",
+    "eslint-plugin-import": "^2.3.0",
+    "eslint-plugin-jsx-a11y": "^5.0.3",
+    "eslint-plugin-node": "^5.0.0",
     "eslint-plugin-promise": "^3.5.0",
-    "eslint-plugin-react": "^6.10.3",
-    "eslint-plugin-standard": "^2.1.1",
+    "eslint-plugin-react": "^7.0.1",
+    "eslint-plugin-standard": "^3.0.1",
     "node-sass": "^3.13.0",
-    "postcss": "^5.2.16",
-    "postcss-cli": "2.6.0",
-    "webpack": "^2.3.1",
+    "postcss": "^6.0.1",
+    "postcss-cli": "^4.1.0",
+    "webpack": "^2.6.1",
     "webpack-merge": "^4.1.0"
   }
 }
