@@ -1,6 +1,7 @@
-import React, { PropTypes } from 'react'
+import React from 'react'
 import { connect } from 'react-redux'
 import { ipcRenderer } from 'electron'
+import PropTypes from 'prop-types'
 
 import * as timerActions from '../../actions/timerwindow/timer'
 import { zeroPad, calcTotalSeconds, convertTimeToNumber } from '../../utils/timeConverter'
