@@ -1,9 +1,8 @@
-import createLogger from 'redux-logger'
+import logger from 'redux-logger'
 import { createStore, applyMiddleware, compose } from 'redux'
 
 import timerReducer from '../reducers/timerwindow'
 
-const logger = createLogger()
 const finalCreateStore = compose(
   applyMiddleware(logger),
 )(createStore)
